relations/requirements.txt
click==7.1.2
psycopg2-binary==2.8.5
ujson==2.0.3
